refactor authentication from server to services
