+++ server/routers/admin.js
@@ -2,9 +2,9 @@ var express = require('express');
 var router = express.Router();
 var pg = require('../../psql-database');
 var bodyParser = require('body-parser');
-var auth = require('../auth.js');
+var services = require('../services');
 
-var ensureAuthorized = auth.ensureAuth;
+var ensureAuthorized = services.ensureAuth;
 
 router.use(express.static(__dirname + '/../../react-client/dist'));
 router.use(bodyParser.json());
