+++ services/config/example.config.js
@@ -1,0 +1,10 @@
+// Remember to set your service keys below
+// Save this file as 'config.js' to use locally
+
+// In Heroku set these as environment variables
+// https://devcenter.heroku.com/articles/config-vars
+module.exports = {
+  STRIPE_PUBLISHABLE_KEY = '',
+  STRIPE_SECRET_KEY = '',
+  PASSPORT_JWT_SECRETORKEY = ''
+};
